requirements.txt
anyio==4.12.1
appnope==0.1.4
asttokens==3.0.1
certifi==2026.2.25
comm==0.2.3
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
docx==0.2.4
executing==2.2.1
fonttools==4.56.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ipykernel==7.2.0
ipython==9.11.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
joblib==1.5.3
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.4.8
lxml==5.3.0
matplotlib==3.10.0
matplotlib-inline==0.2.1
mpmath==1.3.0
nest-asyncio==1.6.0
numpy==2.2.3
packaging==24.2
pandas==3.0.1
parso==0.8.6
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.9.4
prompt_toolkit==3.0.52
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-telegram-bot==22.7
pyzmq==27.1.0
scikit-learn==1.8.0
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
stack-data==0.6.3
sympy==1.13.3
threadpoolctl==3.6.0
tornado==6.5.4
traitlets==5.14.3
typing_extensions==4.15.0
wcwidth==0.6.0
